責務分離
